IO_Blind_Move Product_code vira long para se adequar aos outros processos @#$ T_IO_Serial_Process_Download_Move Metodo para gerar objeto IO_Blind_Move da Response @#$ WS_IO_Serial_Process_Download.java Tratamento da Resposta para gerar IO_BLind_Move e passar PK para act chamadora @#$ IO_Blind_Move_Sql_002 Gera auto-incremento de blind_tmp @#$ IO_Blind_Move_Sql_004 Seleciona Movimentação unica @#$ Act052_Main_Presenter Passa PK de movimentação e direciona para Act058 @#$ Act055_Main_Presenter Correcao de chave para passar prefixo @#$ Act058_Main Recuperação de dados do Frag_Move_Create separando a origem entre move ou move_planned @#$

+++ app/src/main/java/com/namoadigital/prj001/model/IO_Blind_Move.java
@@ -8,7 +8,7 @@ public class IO_Blind_Move {
     private int blind_tmp;
     private Integer blind_prefix;
     private Integer blind_code;
-    private int product_code;
+    private long product_code;
     private int serial_code;
     private String serial_id;
     private int site_code;
@@ -61,11 +61,11 @@ public class IO_Blind_Move {
         this.blind_code = blind_code;
     }
 
-    public int getProduct_code() {
+    public long getProduct_code() {
         return product_code;
     }
 
-    public void setProduct_code(int product_code) {
+    public void setProduct_code(long product_code) {
         this.product_code = product_code;
     }
